+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [v5.3.0] - 2025-08-26
+### Added
+- [#118] Add setup timeout variables to ces terraform module
+
+### Changed
+- [#116] Update setup version to latest to avoid no longer available bitnami images
+- [#115] Updated documentation for ecosystem certificate
 
 ## [v5.2.0] - 2025-03-21
 ### Added